Implementing TenantInterecptor
It is responsible to block requests based on the value of the
X-Tenant-Id header. If the header is missing or if there is none tenant
associated with the value an error response is sent back to the client.
Otherwise the tenant informed is set in the static context and the
response chain continues.